+++ tests/test_03_cli.py
@@ -13,7 +13,7 @@ from conftest import (require_compression, setup_testdata, testdata_items,
 
 
 # Setup a directory with some test data to be put into an archive.
-# Make sure that we have all kind if different things in there.
+# Make sure that we have all kind of different things in there.
 testdata = {
     "dirs": [
         (Path("base"), 0o755),